Create WebDocumentFactory and refactor Crawler
