+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/szechenyilibrary/Crawler.java
@@ -1,10 +1,7 @@
 package com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary;
 
 import com.reka.lakatos.searchbyisbn.crawler.BookCrawler;
-import com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.session.Session;
-import com.reka.lakatos.searchbyisbn.crawler.szechenyilibrary.session.SessionManager;
 import com.reka.lakatos.searchbyisbn.document.Book;
-import com.reka.lakatos.searchbyisbn.webdocument.WebClient;
 import com.reka.lakatos.searchbyisbn.webdocument.WebDocument;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -18,14 +15,9 @@ import java.util.List;
 @RequiredArgsConstructor
 @ConditionalOnProperty(name = "crawler.book-crawler", havingValue = "szechenyi")
 public class Crawler implements BookCrawler {
-
-    private final SessionManager sessionManager;
-    private final SessionActivationChecker sessionActivationChecker;
     private final DocumentReader reader;
-    private final WebClient webClient;
+    private final WebDocumentFactory documentFactory;
 
-    private String currentServerUrl;
-    private String currentServerSessionId;
 
     @Override
     public List<Book> getNextBooks() {
@@ -34,24 +26,11 @@ public class Crawler implements BookCrawler {
     }
 
     private List<String> getBookListLinks() {
-        setCurrentServerUrlAndCurrentServerSessionId();
-
-        WebDocument webDocument = webClient.sendPostRequest(
-                currentServerUrl,
-                "SESSION_ID="
-                        + currentServerSessionId
-                        + "&lv_action=LV_Scan&new_scan=-1&SCAN_TERM=978-963-0&SCAN_USE=BN&SCAN_PREFERRED_POSITION_IN_RESPONSE=2&SCAN_NUMBER_OF_TERMS_REQUESTED=10&SCAN_STEP_SIZE=0");
-
-        System.out.println(sessionActivationChecker.isSessionActive(webDocument));
+        WebDocument webDocument = documentFactory.getSearchingResult();
 
         List<String> elements = reader.getBookPropertiesPageLinks(webDocument);
         elements.forEach(System.out::println);
         return null;
     }
 
-    private void setCurrentServerUrlAndCurrentServerSessionId() {
-        Session session = sessionManager.getActiveSession();
-        currentServerUrl = session.getServerUrl() + "lv_scan.html";
-        currentServerSessionId = session.getId();
-    }
 }
